Fix PR feedback Move announcements code/logic to seperate/ its own folder Move and refactor email service to seperate/ its own folder
