+++ backend/ng/announcements/__init__.py
@@ -1,0 +1,3 @@
+"""
+Announcement Module
+"""
