+++ backend/ng/announcements/controllers/__init__.py
@@ -1,15 +1,18 @@
 """
-Admin action controllers for notifications
+Announcement controllers
 """
 
+from .get_announcements import (
+    get_active_announcements,
+)
 from .manage_announcements import (
     send_announcement,
     send_event_announcement,
     get_all_announcements,
 )
 
-
 __all__ = [
+    "get_active_announcements",
     "send_announcement",
     "send_event_announcement",
     "get_all_announcements",
