+++ backend/ng/announcements/models/__init__.py
@@ -1,0 +1,10 @@
+"""
+Announcement models
+"""
+
+from .Announcement import Announcement, AnnouncementType
+
+__all__ = [
+    "Announcement",
+    "AnnouncementType",
+]
