+++ backend/ng/announcements/services/__init__.py
@@ -1,0 +1,10 @@
+"""
+Announcement services
+"""
+
+from .announcement_service import AnnouncementService
+
+
+__all__ = [
+    "AnnouncementService",
+]
