+++ backend/ng/announcements/tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Announcement Testing Module
+"""
