+++ backend/ng/emails/__init__.py
@@ -1,0 +1,3 @@
+"""
+Email Notifications Module
+"""
