+++ backend/ng/emails/services/__init__.py
@@ -1,0 +1,11 @@
+"""
+Email services
+"""
+
+from .email_notification_service import TicketEmailService
+from .email_sender import get_email_service
+
+__all__ = [
+    "TicketEmailService",
+    "get_email_service",
+]
