+++ backend/ng/emails/templates/new_ticket.html
@@ -1,0 +1,42 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <style>
+        body { font-family: Arial, sans-serif; line-height: 1.6; color: #333; }
+        .container { max-width: 600px; margin: 0 auto; padding: 20px; }
+        .ticket-info { background: #f5f5f5; padding: 15px; border-radius: 5px; margin: 20px 0; }
+        .button { background: #007bff; color: white; padding: 10px 20px; text-decoration: none; border-radius: 5px; display: inline-block; }
+        .footer { margin-top: 30px; padding-top: 20px; border-top: 1px solid #ddd; font-size: 12px; color: #666; }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <h2>New Support Ticket Created</h2>
+
+        <div class="ticket-info">
+            <p><strong>Ticket #{{ ticket.id }}</strong></p>
+            <p><strong>Subject:</strong> {{ ticket.subject }}</p>
+            <p><strong>From:</strong> {{ ticket.author_name or 'Unknown User' }}</p>
+            <p><strong>Created:</strong> {{ ticket.opened_timestamp|format_datetime }}</p>
+            {% if ticket.event_name %}
+            <p><strong>Event:</strong> {{ ticket.event_name }}</p>
+            {% endif %}
+            {% if ticket.team_name %}
+            <p><strong>Team:</strong> {{ ticket.team_name }}</p>
+            {% endif %}
+            {% if ticket.challenge_name %}
+            <p><strong>Challenge:</strong> {{ ticket.challenge_name }}</p>
+            {% endif %}
+        </div>
+
+        <p>
+            <a href="{{ ticket_url }}" class="button" style="color: #ffffff; text-decoration: none;">View Ticket</a>
+        </p>
+
+        <div class="footer">
+            <p>This email was sent by the CTF Support System.</p>
+        </div>
+    </div>
+</body>
+</html>
