+++ backend/ng/emails/tests/__init__.py
@@ -1,0 +1,5 @@
+"""
+Email Service Testing module
+NOTE: _test_email_send.py tests sending a real email,
+test_email_integration.py mocks the sending but tests the business logic
+"""
