--- backend/ng/notifications/controllers/all_actions/__init__.py
@@ -1,10 +1,0 @@
-"""
-Shared controllers for notifications
-"""
-
-from .get_announcements import get_active_announcements
-
-
-__all__ = [
-    "get_active_announcements",
-]
