+++ backend/ng/notifications/routes/__init__.py
@@ -2,11 +2,9 @@
 Notification routes package
 """
 
-from .user_routes import notifications_user_namespace
-from .admin_routes import notifications_admin_namespace
+from .routes import notifications_namespace
 
 
 __all__ = [
-    "notifications_user_namespace",
-    "notifications_admin_namespace",
+    "notifications_namespace",
 ]
