+++ backend/ng/notifications/services/__init__.py
@@ -1,15 +1,9 @@
 """
-Notification Service Package
+Notification services
 """
 
-from .email_templates import TicketEmailTemplates
 from .notification_service import NotificationService
-from .email_service import get_email_service, AWSEmailService
-
 
 __all__ = [
     "NotificationService",
-    "get_email_service",
-    "AWSEmailService",
-    "TicketEmailTemplates",
 ]
